% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conflicts.R
\name{r2academy_conflicts}
\alias{r2academy_conflicts}
\title{Conflicts between the r2academy and other packages}
\usage{
r2academy_conflicts()
}
\description{
This function lists all the conflicts between packages in the r2academy
and other packages that you have loaded.
}
\examples{
r2academy_conflicts()
}
