% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{r2academy_deps}
\alias{r2academy_deps}
\title{List all r2academy dependencies}
\usage{
r2academy_deps(recursive = FALSE)
}
\arguments{
\item{recursive}{If \code{TRUE}, will also list all dependencies of
r2academy packages.}
}
\description{
List all r2academy dependencies
}
