% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{r2academy_packages}
\alias{r2academy_packages}
\title{List all packages in the r2academy}
\usage{
r2academy_packages(include_self = TRUE)
}
\arguments{
\item{include_self}{Include r2academy in the list?}
}
\description{
List all packages in the r2academy
}
\examples{
r2academy_packages()
}
