% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{r2academy_update}
\alias{r2academy_update}
\title{Update r2academy packages}
\usage{
r2academy_update(recursive = FALSE)
}
\arguments{
\item{recursive}{If \code{TRUE}, will also check all dependencies of
r2academy packages.}
}
\description{
This will check to see if all r2academy packages (and optionally, their
dependencies) are up-to-date, and will install after an interactive
confirmation.
}
\examples{
\dontrun{
r2academy_update()
}
}
